feat: abstract utility function
